Icelandic QWERTY Layout: Add ability to write Icelandic quotation marks („“)
This has been part of the Icelandic Dvorak & Icelandic Macintosh layouts
that I added for a long time now, I *thought* I'd submitted a patch to
add this to the main QWERTY layout, but evidently I was wrong.

These quotation marks are the official ones in Icelandic, but the
Windows layout doesn't have them (it's Alt + 0132 & Alt + 0148,
respectively), there's some archane Unicode entry think on Ubuntu
apparently that doesn't work for me on Debian+GNOME.

Due to this patch you can no longer write ð and Ð with altgr combos on
the "d" button, but I think that's acceptable since it has its own
dedicated letter on the keyboard in this layout, and the only reason
this was possible was because it inherited it from the basic latin
layout.

Why use the "d" button? No reason other than I was scouring the keyboard
for keys that were either the same under the altgr combos as the
non-altgr versions (e.g. "n" is such a letter, but there's not many), or
were redundant to other letters, "d" was the closest to "2" of those and
far away from the altgr button, so it's easy to type it, it's also close
to other common special altgr combos like "@" which is produced by
pressing altgr+q.

The /r/Iceland Reddit thread that prompted me to finally write this
patch:
https://www.reddit.com/r/Iceland/comments/40nwr2/einhverjir_linux_menn_h%C3%A9r/